+++ src/hackingBuddyGPT/usecases/web_api_testing/simple_openapi_documentation.py
@@ -1,4 +1,3 @@
-import time
 from dataclasses import dataclass, field
 from typing import List, Any, Union, Dict
 
@@ -7,15 +6,14 @@ from openai.types.chat import ChatCompletionMessageParam, ChatCompletionMessage
 from rich.panel import Panel
 
 from hackingBuddyGPT.capabilities import Capability
-from hackingBuddyGPT.capabilities.capability import capabilities_to_action_model
 from hackingBuddyGPT.capabilities.http_request import HTTPRequest
 from hackingBuddyGPT.capabilities.record_note import RecordNote
 from hackingBuddyGPT.usecases.common_patterns import RoundBasedUseCase
-from hackingBuddyGPT.usecases.web_api_testing.documentation_handler import DocumentationHandler
-from hackingBuddyGPT.usecases.web_api_testing.llm_handler import LLMHandler
+from hackingBuddyGPT.usecases.web_api_testing.utils.documentation_handler import DocumentationHandler
+from hackingBuddyGPT.usecases.web_api_testing.utils.llm_handler import LLMHandler
 from hackingBuddyGPT.usecases.web_api_testing.prompt_engineer import PromptEngineer, PromptStrategy
-from hackingBuddyGPT.usecases.web_api_testing.response_handler import ResponseHandler
-from hackingBuddyGPT.utils import LLMResult, tool_message, ui
+from hackingBuddyGPT.usecases.web_api_testing.utils.response_handler import ResponseHandler
+from hackingBuddyGPT.utils import tool_message
 from hackingBuddyGPT.utils.configurable import parameter
 from hackingBuddyGPT.utils.openai.openai_lib import OpenAILib
 from hackingBuddyGPT.usecases.base import use_case
