+++ src/hackingBuddyGPT/usecases/web_api_testing/utils/openapi_converter.py
